# spec/openapi.yaml
openapi: 3.1.0
info:
  title: GPS-Mock Engine — REST API
  version: 0.1.0
  description: >
    HTTP contract of the Go engine (v3). REST is intentionally minimal — just
    status/health plus the two simplest location actions. Every other action
    (routing, sequences, patrol, favorites, history, settings, driver switch,
    device info, admin actions...) is WebSocket-only; see asyncapi.yaml for
    those. This split mirrors the actual server implementation
    (engine/internal/server/server.go): don't add a REST path here unless a
    handler for it actually exists, and don't remove a WS action from
    asyncapi.yaml without removing its dispatch case too.
servers:
  - url: http://{host}:{port}
    variables:
      host: { default: localhost }
      port: { default: "8080" }
paths:
...
  /api/location/set:
    post:
      summary: Inject a spoofed GPS position
      operationId: setLocation
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SetLocation" }
      responses:
        "200": { description: Result, content: { application/json: { schema: { $ref: "#/components/schemas/OpResult" } } } }
...
components:
...
  schemas:
    OpResult:
      type: object
      properties:
        success: { type: boolean }
        error: { type: string }
        ignored: { type: boolean }
      required: [success]
...
    Coords:
      type: object
      properties:
        lat: { type: number }
        lon: { type: number }
        name: { type: string }
      required: [lat, lon]
    SetLocation:
      allOf:
        - $ref: "#/components/schemas/Coords"
...